requirements.txt
Flask==2.3.3
flask-cors==3.1.1
google-generativeai==0.0.6
python-dotenv==1.0.0
requests==2.31.0
Jinja2==3.1.2
MarkupSafe==3.0.2
